Refactor folder structure and rewrite Event API
This commit organizes the code so the future Server and Client APIs are decoupled.

It also rewrites the server Event API, so instead of using a chaining API for adding fields to the event the constructor takes variadic arguments. Some fields are also rewritten as type definitions for their underlying base types instead of structs for easier use. For this to be done, the SingleLine optimization was removed. If the Chunk Scanner is deemed to be a bottleneck, a way to add the optimization back without breaking the API will be considered.