Recover from panics on context done and log write errors.
As the handler may be stopped externally (for example on Shutdown), it is impossible for the requests to know when this happens and not send the disconnect command to the hub. This means that when the hub is closed the disconnect goroutine panics. Recovering from the panic is enough to solve this issue, as it doesn't occur elsewhere.
Also add an option to handle possible write errors in a custom manner. While they might be a rare occurence it may prove vital to know why they occur.