Improve docs, create context setters, handle new connections to closed handler
This commit creates type-safe context value setters for last event ID and topics, as recommended by the standard library. It also creates some internal getter functions for those values, to clean up the handler code.
The situation where a new request arrives, but the handler is closed is also taken care of. Without aborting the connection, the request would hang indefinitely, as the connection request would never be received.
Other changes include removing the OnConnect and OnDisconnect handlers (continued from last commit), expanding the documentation and rearranging the code a bit, so it looks cleaner.