Write space after colon and remove fieldWriter allocation
Given that the protocol requires only the first space after the colon in a field line to be removed, we always write one so the data won't get truncated if it starts with spaces.
The field writer is also refactored so heap allocations are avoided for the fieldWriter struct itself.