Decouple message broker and HTTP handler
This is a drastic API change that creates immense opportunities for customization. By not forcing the channel-based message broker and a single HTTP handler implementation, we can allow for custom brokers and handlers that suit the use-case of every person.
Some examples include using Redis pub-sub/Apache Kafka/RabbitMQ as message brokers, HTTP handlers that deal with authentication or other custom logic. We are no longer bound to context values for Last-Event-ID or topics, as they are now simply a parameter for the provider's Subscribe method.
This also allows using the same infrastructure with other HTTP frameworks by simply implementing a new Connection type.