Improve replay provider API and make Joe more reliable
This commit ensures Joe closes the subscribers even if he panics.
The replay provider interface does not require returning errors for Put and Replay, as they weren't handled anyway. The GC error has now a useful meaning. Finally, the documentation is more complete, as the requirements for a replay provider are fully stated.