Implement TextUnmarshaler for Event, create event marshaler interface
The event marshaler interface allows users of the library to create custom types that can represent themselves as events. This way data that can be sent as an event can be represented in a custom manner, which allows marshaling to other formats (JSON, XML, binary etc.) and saving in different locations (disk, database) in the most optimal way, which wouldn't be possible using the MarshalText and UnmarshalText implementations on the Event alone.
The usage of the interface is also shown in the complex example.