Merge field types under a single type
This change removes the indirection created by using interfaces for
fields, which should improve memory usage and performance. It also reduces
code duplication to a minimum.
Field deduplication for events is also removed, as clients are supposed
to only use the last value, so it would be redundant to dedupe them.
This change removes the event.ID type, which may have provided useful
context instead of strings for working with IDs in servers/providers.
Consider an alternative way to achieve this.