Fix parsing events without newline
The spec says that if an event doesn't end in a
newline it is not dispatched. This wasn't the case in
our implementation.
ByteParser now returns a parser error if it finds a
field that doesn't end in a newline, so ReaderParser
stops scanning and the event isn't dispatched anymore.
This can occur only when a reader reaches EOF, so this
means that it is impossible for transmissions to be
ended prematurely.
The change is reflected also when unmarshaling server
events from text, as the error is now returned if the
byte slice doesn't end in a newline.