Defer headers flush to first send
Sending the upgrade headers only before sending the
first message allows for HTTP handlers to exit early
after calling Upgrade. This gives greater flexibility
in creating custom handlers. It also allows handlers
to try and connect to providers after upgrading, which
in case the upgrade fails the request does not even
reach the provider. This was not possible before,
which in turn might have caused subscriptions to
providers to be established, only for them to be
canceled because the handler wouldn't support
streaming.