Improve ReplayProvider interface, use callbacks
Replaying is now aborted if the subscription callback
returns an error. This was not possible in the
channel-based approach used before, which in turn may
have resulted in a deadlock: on error, the message
receiver would return, but the replay provider would
keep sending messages.
We also remove the requirement for replay providers to
implement GC. Instead, we create another interface,
and providers have to check if the given replay
provider implements that interface, so they can GC
when necessary. This removes the need for no-op GCs
to return an error, so providers can then stop GC-ing.
Some tests are buggy after this change, so they are
skipped.