server: Reduce allocations
To discriminate between a write error and a subscribe error, we would
wrap the error from Write in a struct. This caused a heap allocation. We
now use a flag that indicates whether an error has occurred, which we
check inside ServeHTTP when Subscribe returns.
We also don't allocate a topics slice every time.